+++ docs/release-notes.rst
@@ -23,6 +23,12 @@ Common Data Service Release Notes
 The client and server are released together.  The client is deployed as a jar file to a
 Maven Nexus repository. The server is deployed as a Docker image to a Docker registry.
 
+Version 1.16.1, 2 Aug 2018
+--------------------------
+
+* Store API token encrypted, not hashed (ACUMOS-1487)
+* Requires database schema version 1.16
+
 Version 1.16.0, 24 July 2018
 ----------------------------